FlightPoints - first messy version
For now i only figured and partilly fixed tooltip on the destination node.
However data seems to be different from the WoTLK Classic. And would need to be completely rewritten.
Also i have problem with debug string as it is currently unable to get hops FPs in between start and end nodes correctly...

And it seems that my routestring is bad too. As it is for some reason adds un-needed zeroes between nodes
For example: FP from Orgrimmar to Valormok, no hop in between them, Orgrimmar added twice and a lot of un-necessary zeroes in between.
["0.63:0.56:0.00:0.00:0.63:0.64"] =  -- Orgrimmar, Orgrimmar, Valormok