FlightPaths - DB, cleanup
NodeOnClick - "Mimic" the PLAYER_CONTROL_GAINED event, by using repeating timer, while player is on flight (taxi) .
The repeating loop helps us to hide flight bar when player is out of taxi.
This loops is majorly needed for a player to be able to send reports to the author, for if current flight path is not in database, or flight time is not right.
ReportMissingTimer - there is little problem. I made it to not check for the "hops" to print the debug. So for example, if flight is from Dalaran to Dun Niffelem, it may go different ways. Through k3 and through zuldrak. But my print will always say Dalaran, Dun Niffelem. So it could be confusing. But at least i hope all the routeString are fine.
DB - Updated DB from WOTLK Classic