FlightTimer - improve, fix
Test fix for such events as summon, BG/Arena entry to stop the bar and hide editbox (report frame).
DebugString - now adds hops count (to avoid confusion i talked already about)
To Do: Need to fix editbox for 3.3.5, it is right now not copyable and look of it is bad. Can take RecentChatWindow as a reference, as i already fixed that one.