media: fix incorrect duration
since we backported the addon from WoTLK classic, some tracks, have incorrect duration (usually +1s)

which results in wrong ending of the track, it starts all over again for 1 sec.

so this is the process of fixing it.

we getting EpicMusicPlayer data compare to ours and fix our DB durations for incorrect tracks data.